backend_fastapi/pyproject.toml
[project]
name = "backend_fastapi"
version = "0.1.0"
description = "A production-ready FastAPI with uv and loguru."
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "pydantic-settings>=2.3.4",
    "pydantic[email]>=2.0.0",
    "loguru>=0.7.2",
    "jinja2>=3.1.2",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "motor>=3.3.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==3.2.2",
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "cryptography>=41.0.0",
    "email-validator>=2.0.0",
]

[project.scripts]
dev = "app.main:run_dev_server"
prod = "app.main:run_prod_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include the 'app' directory as a top-level package.
packages = ["app"]
